Add a prototype building shader
The first goal for this project is to create a singular building.
That's it, just a singular building.

In my original Neon City project, each building was composed of about
three components: the street level color, the dark building color, and
the window color. I originally constructed different meshes for each
of these, but switched to shaders for performance reasons.

So, I've added in a prototype shader that sort-of replicates how that
other shader worked. Only difference is that this new shader uses
tri-planar textures, which should make working with oddly-shaped
buildings easier.

Now, the only problem is that Godot lacks any way for us to generate
textures, which means I'll have to make them by hand...