Add a script for window-building textures
In the original Neon City generator, I dynamically created textures at
runtime using Java's native AWT package. It was really neat!
Unfortunately, Godot doesn't have the same functionality. We can't
generate anything at runtime, so we need to provide purely prebaked
assets.

That's a bit of the problem - I was creating the textures dynamically
both so the city would be unique each time, but also so I DIDN'T HAVE
TO BECAUSE I AM NOT AN ARTIST. So I still needed something to generate
the textures for me, because I didn't feel comfortable doing that
myself.

So, I've cooked up a Bash script to do it for me! It runs using a
pretty simple algorithm - it glues together different types of images
into rows, which it then glues into a larger texture image. It does
this using the Image Magick tool, which is quickly becoming a favorite
for en-masse image editing.

Now I only have one problem to solve - it's REALLY SLOW!