Speed up gluing together images with an array
Previously, we were excuting Image Magick operations one at a time,
attaching the images to the end of a working-row. This was incredibly
slow. I had originally thought we had to do it like this since we were
changing up the images we were using constantly.

However, I realized that we could build up a list of the images to use
and then execute the gluing mechanism all at once. Crazy enough, it
actually worked!

It now takes ~20 seconds instead of however long it took before
(admittedly I wan't keeping track). I'd consider that acceptable for
what's basically part of our asset pipeline - speed isn't exactly of
the essence.