Add choices for different types of windows
Previously, we only had one type of image we could choose from. I was
pretty happy with that, but like the vertical/horizontal construction,
I had an idea to improve it. We could use different types of images to
create buildings that "feel" different.

So, I've added in new window types and a menu for choosing the type of
window. This allows for 12 different window textures, which I think is
pretty good!