Lay the groundwork for auto generating buildings
We're at cross-purposes here for what we want to do with our textures:
we want the fade/gradient texture to be locked in place, but we want
the window texture to be random to each building and each SIDE of the
building.

Fortunately, Godot allows for two sets of UV textures/coordinates: UV
and UV2. So, we're gonna use the UV coordinates as the 'locked'
coordinates and we're gonna use the UV2 as the 'random' coordinates.

However, to be random to each building, we need to directly build in
the UV/UV2 coordinates by hand. So, I've copied some code from my
ongoing 'fictional-guacamole' project. It works, but to be honest,
it's a bit of a mess. It'll take some time to REALLY sort out how best
to do this...