Generate a basic building
Now that we've got some code for generating buildings, we need to do
the actual generation. I'm taking a real slap-it-together approach,
but I think it will work in the long run (with the caveat of some
refinement).

Firstly, the auto-building now generates a building by cell / window
count, instead of by world-unit size. We do this using a UV_SIZE
constant. Turns out, in Godot, the pixel-per-world-unit is
1 / texture size. So I substituted the number of windows - 32 - to get
the appropriate world-unit-per-window size.

There does seem to be something off about the UV of the windows. When
a side is odd, it needs to be shifted by half of a window. If it's
even, it is fine.

I also need to figure out a solution for the roofs. The ideal camera
angle is actually going to be aiming upwards so IDEALLY we won't
actually see the roofs. In the old neon-city generator this was
handled by using vertex color - I wonder if it would be worth it to
return to that instead of texture...