Adjust UV shift for side parity
I don't exactly understand how textures work in Godot. It seems
they're tied to a world unit, so a texture will always occupy one
world unit regardless of the dimensionsions of said texture. It also
seems, at least for SurfaceTool-generated Meshes, that the texture is
centered in the middle - I THINK.

The exact mechanics don't matter as much here as the result: if an
AutoBuilding had an odd side, then the windows on that side would have
bisected windows at the edges. That had something to do with the
texture centering.

Anyway, the solution was to shift the texture by half a window IF that
texture was odd. In code, I just leveraged the parity to turn the
shift on-and-off - allows me to skip using if statements.