Randomize UV Shift for cells/windows per-face
Previously, all of the building faces were using the world aligned UV
values (save the odd-even correction). The problem was that meant
every single face of the building was the same, except for some
incidental mirroring.

So, I've added a new function to calculate the cell-based UV shift
with some random thrown in for good measure. I moved the odd-even
correction into this function to help simplify everything visually.

While the result works, I'm unsure how much I like it. I think the
window texture is too small - we need a larger texture, which means
we'll need a larger street fade texture. Though I'm growing less and
less crazy about the dithering. The smooth colors of the original Neon
City Generator was one of that project's greatest features. I need to
try and recapture that somehow...