Add a prototype 'Light' spatial shader
Every iteration of this project, every time it's crossed my mind, I've
said out loud, to myself: "I wish I could just directly grab the light
and shadow colors and manipulate those directly". I then assumed that
this was impossible and gave up.

Turns out it IS possible. At least in Godot. I assume that it exists
in other places. It's just a bit poorly documented. I wasn't exactly
sure how to do it, so I built an example I found on the internet using
the VisualShader code.

It seeeeeeems to work pretty good. The colors aren't one-for-one, and
there's a lot of playing around we can do. BUT - it's a start!