Add windows to the light-shader
Experimenting some more with the light-based shader, I TRIED to 'cap'
the light that was affecting the surface. The idea was, by clamping
the energy value extracted from the light_color input, I could stop
the color from mutating too much - i.e. going too white.

It didn't work. I'm sure I could use MATH to figure it out (and maybe
I will at some point) but for right now this looks good enough!

So, to kind of 'christen' the light shader, I've added in the windows.
Unlike previous versions, this doesn't happen in the fragment shader,
because that would be overwritten by the light color. Here, it happens
in the light shader. That way, the windows appear unshaded, no matter
the background color!