Convert AutoBuilding to use shaders exclusively
After experimenting with the light shader, I've decided that it's JUST
the right thing for this project. I'm really happy with it, I feel
like it offers more utility than that other shader.

The old shader - while it WAS cool to have textures - was rigid and
inflexible. Just really really hard to work with, honestly. There was
no way to control the size of the fade without manually rebuilding the
texture. Plus, the fade texture HAD to be the same size as the window
texture, constraining the size that file could be.

Since the new lighting shader uses lights to produce the color-fade,
we don't need a texture just for the fade. This all means we can scale
the window texture up and down as we please and use as many colored
lights as we want. WOOP!

Considering the benefits of the lighting-shader, I deleted the old
shaders and retrofitted the AutoBuilding for the new shader. This
simplified our code a lot, since we now depend on UV instead of UV2.
Overall, I really like this new direction! Now, to do something about
these windows...