Use partial transparencies to create "dim" windows
I was happy with how the shaders were working, but I wasn't happy with
the windows. I could see (or felt like I was seeing) repeating
patterns in the windows. I also felt like the texture was somewhat
bland. The original Neon City Generator had dim windows in addition to
the regular windows, so I decided to implement those.

This had always been the plan - that's why the script windows were
called 'A' to differentiate them from the 'B' semi-transparent series.
I more or less lifted the code from the original Neon City project,
and it already looks much better!

I also increased the window-texture size to 64x64, because that what
the original Neon City used. I think. Whether that's an accurate
conception or not doesn't matter too much, because I think the 64x64
texture looks much better. It's harder to notice any distinct pattern.

Now for a different problem: our shader only does one light at a time!
I don't know why that's happening. Multiple lights are critical for
this project, so we'll need to sort this issue out!