Pack together AutoBuilding into scenes
The AutoBuilding had been just sorta floating around in the main
scene. It was time to really formalize that into it's own separate
scene.

So, the AutoBuilding has become the AutoTower - since it's just a
square-rectanguloid tower. This will be the simplest form of building.
The idea is that we'll utilize this for constructing more complicated
buildings.

To go with that, the AutoBuilding was modified to allow toggling of
the automatic-build-on-entry feature, as well as to to toggle whether
the building has windows or not.

The FullServiceBuilding class is meant to be the more "fully featured"
scene. Eventually (if we see fit), this scene will include building
warning lights and spires/antennaes. For right now, it's just divided
into the Base and the Main Tower. Who knows! Maybe that will be
enough. Now, to work on the auto-scrolling...