Add an auto-scrolling "sled"
While the building is the base unit of this project, we needed a way
to semantically divide up the scrolling city that was more manageable,
because the city now consists of lights AND buildings.

That's where the new ScrollSled comes in. This is meant to be the real
"meat" of the city. This is the unit that we'll spawn in, have scroll
across the screen, and then delete again.

To help make some constants more accessible, I extracted them from the
AutoTower to a new utility script - GlobalRef. That way we didn't have
to keep reaching down into the AutoTower scene to get what we needed.

The only problem I'm having now is deciding what should be controlled
by the sled versus what should be controlled externally. For example,
should there be a function to set the dimensions of the building in
the ScrollSled? Or should we reach down and do it manually? And who
handles the building? We'll have to play around with it.