Spawn and organize sleds using lanes
With the scroll sled, we have an effective way of spawning a building
and a light, and then making them scroll together. A good way of
bundling things.

But we needed something to manage those bundles - and that's where the
SledLanes came in. The lane objects control the random aspects of
their respective sleds. They handle spawning the sleds, ensuring they
are appropriately spaced out and randomized within the lane's
variables.

This necessitated changes to the ScrollSled scene - specifically
because the SledLanes needed to modify the ScrollSled's building and
other such elements. We'll probably modify this more as time goes
on...