Give the lane a debug mesh to assist placement
I was having trouble arranging the different lanes because they have
large potential areas but the lanes in no way reflected that. It made
it difficult to predict the outcome of placing a lane in a given
location without rigorous guess-and-check.

So, I've added a mesh that expands and contracts to a theoretical
maximum-possible-size to help us align our building spawns. This has
allowed me to add even more buildings to the main scene.

The only issue is that the mesh expands to the theoretical maximum.
It fails to show the width or the depth - either the minimum or the
maximum. Most of the time, it seems, the hypotenuse wins out. Which
I probably should have been able to guess, honestly, but that would've
required diverting more (any) of my PERSONAL & CRUCIAL BRAIN RESERVE
to the HORRID ACTIVITY known only as MATH.