Experiment with the camera and density
I'm definitely no artist. Well, maybe I could be, BUT one thing I'm
definitely not is a Director of Photography, nor am I a
cinematographer. So I've really been struggling with how to ground the
camera to appropriately capture the scrolling of the city. I'm also
not interested in trying too hard to become those things. Right now,
anyway. I'm lazy and I like programming more.

My lack of inherent cinematographic talent combined with my inherent
cinematographic apathy meant that I wasn't really happy with how my
city was turning. It didn't feel... dense enough.

Of course, it didn't take to long for me to figure out that I could
solve that with the addition of more buildings. However, it was too
easy to see through the buildings to the bases that I was trying to
hide, so I decreased the FOV to flatten the city out. That gave a
really good impression of depth.

One of my current issues - the mesh I created to help me debug where
the buildings were going to spawn doesn't really work. It always shows
the hypotenuse length (which was predictable). I think I might have
even mentioned this in the last commit? I'm not sure. I really should
fix that...