Improve the debug meshes in SledLane
Previously, the debug mesh we created to demonstrate the properties of
each SledLane consisted of a singular, white, untextured block. The
block was effectively square, each side be equivalent to the max
possible hypotenuse of any possible building. The block was
arbitrarily set to be 5 tall.

This information was not useful. It really wasn't. I needed a way to
see the minimums and maximums. Heights, widths, depths, the whole of
it. And I needed a way to distinguish all the information.

So now there are three separate debug meshes. They are textured to
help gauge scale and are colored to help separate them. The tallest
shows maximum possible height and minimum width-depth. The next shows
minimum height and maximum width-depth. The final is a flat plane to
show the aformentioned hypotenuse-square. That represents the maximum
possible footprint for a building IF it is rotated.

With this visual information I was able to make much more informed
positioning decisions - I moved some of the lanes around and was happy
with the overall result.