Randomize the distance between sleds per lane
Previously, the distance between buildings was fixed, based on an
estimated "max possible distance on x" variable. However, this was
resulting in a very clear pattern emerging. In other words, the city
wasn't looking as random as it needed to.

But there were other problems - as I was arranging the different
layers, I wanted to control how sparse or how dense a layer/lane was.
That way, I could control what parts of the city got more or less
dense based on proximity to the camera.

The solution was to control the spawn rate of the sleds. Technically,
they were already spawning at the maximum possible rate. The only
thing we could really do, then, was decrease the spawn rate by
increasing the distance requirement for spawning.

That's what the new "extra space" configurables for each lane control.
They measure the possible "extra space" in window-units. The RNG is
actually a pure random (white noise) distribution. It seems to work
pretty well!

With all this, I was finally able to experiment with the camera and
building set-up more. Turns out, it's a lot easier to experiment with
camera angles once you've settled on a consistent subject. Seems
obvious, in retrospect. Still, I'm happy with this new layout and
camera angle!