Add a Gaussian-distributed light range value
All of the lights were of a fixed range (size, more or less). This
produced a consistent look across the city. But, it was tooooo
consistent. I wanted more randomness. More variation.

So I added controls for a Gaussian distributed light range. Unlike the
building rotation, this one comes with a both a mean and a variance,
offering us more deliberate controls over how the lights appear per
lane.

So far, this has resulted in a lot more color variation across the
faux city. Just what I wanted!