Add a GUI to live-change the different lights
The ultimate end-goal of this project is to have it only PARTIALLY
dependent on the Godot IDE. I want the user to have some basic metric
of control.

To start, I've added a GUI to allow the user to change our three light
colors on the fly. The new colors slowly trickle in with new
buildings.

I originally considered having the lights change in real time. This
would be feasible with Godot's Group functionality but I was worried
about performance impacts and decided not to.