Add live controls for the building materials
Last time we added live controls for the lights, now we've added live
controls for the materials. This includes selecting the material for
new buildings as well as the uniform colors for the shader params.

It seems to work pretty well. I did make the choice to not live update
existing building's materials. That would've been perfectly feasible,
just... this project's winding down now and I'm really not feeling it
all that much. I'm about ready to pack up and move on.

Also, added a new Dodge material, which I thought would be cool but in
practice just seems to be a bunch of color-jitter nonsense.