Prototype a banded lighting shader
To finish off the project, we're gonna try a dithered shader of some
sort. The plan is to process the light down into separate 'bands'
according to brightness. We can then take the information from these
bands and use them to select the appropriate dithering texture.

To start, I'd though we'd work on the banding. The banding idea itself
is intriguing enough, so I wanted to get that committed first. In this
version, any 'lighter' colors assert themselves over 'darker' colors.
This is done using an if statement - inefficient, but it works for my
needs.

I'm pretty happy so far, now I need to complete the dithering...