Create a multi-colored dither shader
I had an idea for how to implement a different sort of shader: one
where not only are the colors banded instead of a smooth gradient, but
the different levels of banding resulted in entirely different
textures!

The idea was that the different textures would be differing levels of
dither - I initially attempted this with six different banding layers
but this resulted in a weird black-line artifact that I wasn't able to
get rid of. So I fell back to a three-banded shader that used only
three different textures.

The result looks very good when standing still but looks considerably
worse when in movement. However, I'm still pretty happy with the
technique itself so.... good enough? Feels good enough to me!