Add a color-dot script
I had an idea for a new city material, where-in the color is in the
lights on the buildings, which are mostly monotone. The style is
largely inspired by the works of Hiroshi Nagai.

To accomplish that, I need a new script to glue things together
differently. Thus, I redesignated the old window-building script as
the White-Opacity, while my new one is the Color-Dot script.