Create a new dot-oriented shader, remove old stuff
Normally I'm for in-project preservation, but I really wasn't too
happy with how earlier versions of the NEON CITY GENERATOR turned out.
So, when I was creating the new dot-oriented shader, I took a long
look at a lot of different aspects of the project - and I decided to
cut out a lot of the fat.

So, I've deleted all sorts of stuff. I've also, as mentioned, added in
the new dot-window shader. It uses the RGB values of the dots as mix
scalars. As long as the dots are only pure Red, Green, and Blue, then
the shader will pick the corresponding color value.