Use 64x64 colored dots in a variety of styles
Okay, so the 16x16 dots were coming out... blurry. It looked wrong,
especially when the dots and buildings were right up in the camera.
So we needed larger dots - hence, we're using 64x64 dots now.

I made the dots in a variety of sizes too. Since there's only two
styles for the four types of dots, I just went ahead and made 8
textures. Eventually we'll be have live-swapping these textures as
a feature of the city-scroll.