Control the shader and sky options via GUI
This commit finally updates the previously existing GUI to match the
colors and options available for the Dot Window shader-materials.

Some design notes here - one of the things that I disliked about the
previous iteration was that we completely lost track of the default
values once we started changing things. To get around that, I've
instituted this idea of "profiles", where each profile is a hard-coded
group of settings. Changing to a particular profile resets everything
to that basic profile. I want to have some global "custom" profiles,
but I haven't implemented that just yet.

We're also going to have a separate selector just for material type,
in order to freely choose between our different window textures.

Finally, one odd note: we used SpinBox widgets to represent the sky's
different CURVE values. However, those needed to be floats while the
SpinBox only supported ints. To get around that, the SpinBox can go
between values of 1 and 9999 - we then divide this by 10,000 to get
the actual float value we want.