Change building textures on user input
Implemented changing the building textures based on user inputs. I had
previously put off doing this because I thought it would be hard. The
most natural thought I had was one material for each texture, and we'd
then set the parameters on each of those materials EVERY TIME WE
CHANGED a parameter.

Of course, then I realized we only needed to change one parameter: the
textures themselves. We just preload them in our global
MaterialColorControl node, and pass them in as appropriate. That was
MUCH easier.