Add a pause feature to the generator
I was trying to take pictures of the cities I was making but the stupid
buildings weren't standing still (because I made them do that). I
needed a way for them to just STAND STILL, if only for a second.

The solution? A pause button. Now, pressing "space" will pause the
game until the user presses "space" again. Easy!