Add some prototype stars
After viewing the generated city for a while, taking some pictures and
the like, I decided that it needed something more. I have several
ideas, including a moon and some neon signage. The first thing I chose
to work on, though, was a field of stars.

After all, it's just dots, right? How hard could it be? Well, as of
right now, not too hard. It's just a series of dots, which are
actually Quads, instanced via MultiMesh and then sorta scattered
around.

I only have very basic controls on the dots right now. You can set the
height of the starfield, and that's it. Eventually we'll be able to set
the textures, colors, width, scale variance, and number per-subfield.
We'll also have a button to just force a generation pass.

Speaking of subfield: the starfield scene actually consists of three
separate subfields. The idea is that each subfield has a different
mesh instance, corresponding to a different kind/color of star. That's
the idea, anyway.