Flesh out the starfield effect
Previously, we only had in-editor controls for the Starfield's width
and height. I found that unsatisfactory.

We now control the count of stars, per field, as well as the normal
distributed scale across all three fields as a whole. When combined
with changes to the materials, this should allow for good control and
a nice overall aesthetic.