Add GUI items for the Starfield's configurables
This commit adds GUI controls corresponding to the different
configurable items we added to the Starfield class. Changing any of
the values triggers a regeneration of the starfield. If you don't
want to change anything but DO want the stars redrawn, just use the
helpful "Regenerate" button.

The limits I imposed on each of the fields was largely arbitrary; I
may very well go back and change those at a later date.