Make ghost buildings by using Building Color alpha
I was working on something when an idea struck me - what if we phased
out the material of the building and left the windows, creating a sort
of ghost building, or after image of a building. The vague idea of a
building, if you will.

That just required some basic reconfiguring of the existing shader
nodes. Seems to work pretty well - not sure what the utility will be
but I'm excited by the possibilities.