Ensure stars are distributed across entire Starfield
During development, I noticed that stars at smaller scales tended to
cluster towards the upper-middle of the screen. This corresponded with
their origin before the transform.

After some puzzling, I realized the issue: translation and scaling
with MultiMeshes is done using transforms. The scaling was affecting
the translations, regardless of the order the translation & scaling
were applied! Hence, the smaller stars always scaled back to their
point-of-origin. Normally Godot handles this nonsense, but this time
we needed to MATH our way out of this.

The solution was to scale the translation vector by the INVERSE of the
scale we apply to the transform (i.e. 1/scale). That nullifies the
scaling factor and delivers the stars to their correct positions.

Also, I added a seed to the Starfield, specifically when we're running
in-engine. This'll make it easier for us testing wise, and save some
git headaches - the multimesh will no longer change constantly.