Add control of star colors and textures
Okay, so we handled the properties of the Starfield; next up was the
actual stars themselves.

I don't have much to say on the GUI. It was a pretty simple
implementation.

What I should address is the assert_profile method, which grew so
bloated with arguments that I've removed them altogether in favor of
a dictionary. We now declare a default dictionary and modify the
arguments in that before passing them down to the assert_progile
method. This allows us to save some lines... occassionally.

Also, I added some new sprites to the effects directory, based on an
old 1950's travel poster. I think the patterns are referred to as
"starbursts", but I'm not too sure. The specific poster was a very
abstract TWA poster for New York, depicting Times Square (I think).