Avoid alpha around window shapes
I noticed that on larger building-window textures there was an outline
of alpha'd out texture, allowing you to see through the building
around the windows. I presume this was present in smaller
building-window textures, though it must have been too small to
recognize.

This was happening because the building's opacity is a weighted mix
between to color's alpha: a UV sample of the building-window textures
and the building color provided by the user. The weight is the
texture sample's alpha. The idea was this alpha would always either be
fully opaque or fully transparent. However, with fuzzier textures
(such as the dots), this created a value BETWEEN 0 and 1, resulting in
a slightly transparent outline.

This commit updates the shader by FLOOR'ing the texture sample's alpha
before using it as a weight. That seems to have fixed everything.