Tune up the "Velvet Dream" profile
The "Velvet Dream" profile has been the weakest so far - in my
opinion. It feels derivative from Niteflye (probably because they're
both pulled from Hiroshi Nagai's art).

So I was reviewing the artwork that inspired the scheme and found that
I KIND of goofed up the colors. The top of the sky was supposed to be
a much darker, much more rich purple. The ground-horizon, which is
supposed to sort of be the ocean, should be a rich red instead of a
land-blue. And, honestly? The desaturated pink sky-horizon was just
kinda meh. I wanted more saturated colors.

After some playing around, I got a configuration that I liked much
more.