Add a moon
We had stars a-plenty; there was just one final celestial element I
wanted to add: THE MOON!

I more or less covered the core options for the moon: size,
visibility, position, and color. Nothing too crazy going on there.

For most of the profiles I felt like the moon was unecessary, so it's
mostly disabled. But it's something to play with!