Increase the space between superstructures
I was having trouble getting good screenshots because the
superstructure buildings were always too close together. Sometimes you
want that, sometimes you don't - point is, we were always getting it
when you DIDN'T want it. And, how the buildings were spaced out, they
was no opportunity for the buildings to be positioned far enough
apart.

So, I increased the distance in the superstructure lane. Looks good
enough. And I think that, with that, this phase of development has
wrapped. I think I've stretched this fixed-angle generation as far
as I can, or I want to. I don't think I'm explicitly done. I may come
back with a Qodot driven, semi-hand crafted visualizer later on. But
for now, I FEEL done.