Add Qodot to the project
We're going to try a new version of the Neon City generator that
relies on spawning prebuilt buildings. Or city blocks. Or something
like that. Point is, we're gonna start building stuff by hand. And
for that we're gonna need Qodot.

So, I've added Qodot into the project and configured it. So good so
far, I suppose.