Add a set of basic textures for creating buildings
Okay, before we can work in TrenchBroom, we need to have some textures
of our own that we can work with. Those Qodot textures don't really do
us any good at all.

So, I identified four sorts of base level TrenchBroom structures we'll
want. First, we'll have two building materials: one that is earmarked
for horizontal-style windows, and one that will have now windows - in
other words, it will JUST be the base building color.

Then we'll have the street color. In most instances, this should match
the color specified as the Building Base Color, but we'll give users
the authorship to change it IF they want to.

Finally, we have a non-texture: the Light Field. The Light Field will
be a brush-entity, used to a space where 'city lights' should be
randomly populated. This will be done everytime a city block spawns
in.