Add a basic block and a scene to view it
Now that we've got our textures, it's time to actually start building
some blocks.

I decided to start simple. Our blocks will have a 256x256 footprint.
Originally I thought that the streets should be included on the
Southwest (Negative-X, Negative-Z) side of the block. However, I've
decided for now that the buildings should be allowed to occupy the
edge of the footprint.

I have to say, placing the buildings is actually a bit more difficult
than I thought. I know why I automated that process now. We'll have to
investigate the capability for automation and randomization.

I then added a scene to try and view the different blocks. It comes
with a neat little spinning camera, so I'm pretty happy with it so
far.