Randomize blocks via manual UV 'derivatives'
Okay, so I wasn't able to adjust the UV faces like I wanted to via
Qodot because Qodot just builds whatever is in the .map file.

Then I realized something: the .map file is written in a very
particular way. It's deliberately written for easy machine processing,
with the values separated by spaces and presented in a very specific
order. And, I was able to spot which two values represented our UV
shift!

Since Qodot allows us to build at run time, we just need to modify a
file at run time to create a derivative, and then feed it into our
QodotMap node. Presto! A UV shifted building!

Our current technique is to just take in a file and automatically
write a 'derivative' version. We generate the derivative by
processing only those lines that match our building texture. In the
future we'll have stricter control over the randomization, and we'll
probably use some sort of 'caching' to avoid generating unecessary
files.