Upgrade BlockInspector with dyanmic block menu
Previously, the BlockInspector was hard coded to load only a
particular map. For the whole of my Godot experience, this has been
how it had to be. I was always vaguely aware that there was PROBABLY
a better option but, to be quite frank, I'd never really been bothered
by it.

Of course, as we started messing with these dynamic-derivative files,
my eyes opened to the possibility that we could dynamically take stock
of files that we do or don't have.

So, I added a GUI to the BlockInspector. The GUI has an options menu
that is dynamically populated with all of the .map files in the
trenchbroom_maps directory. Pressing the "Go" button will trigger a
new derivative of the selected map to be generated and loaded.

To test this out further, I added two new blocks. Looks good so far!

One odd thing I noticed was that rotating the buildings in TrenchBroom
resulted in ever-so-slightly offset textures. This was resulting in
"bad" windows because we completely ignored the .map file's existing
offset. To counter, we now add in the existing offset to our
derivative offset.