Add camera controls to the BlockInspector
It's only natural that some of the blocks will be bigger than others,
so it was unsurprising that it only took three blocks for us to go off
the screen.

So, I added a GUI to control different aspects of the camera.
Perspective FOV, Orthogonal Size, and camera height. It's perfect - it
allows us to get all different sorts of views on any given block.