Add height properties (and more maps)
Previous iterations of this generator spawned buildings one-at-a-time,
so we were afforded a great amount of control over the heights of the
buildings we spawned. This was perfect since we arranged the buildings
into lanes - farther buildings were supposed to be taller.

Now, however, we're trying to do the handmade approach. We still want
that near-far height dynamic, but there's no immediate method for
determining the bounds of a given block.

So, I've added max_height and min_height properties to each map/block.
Hopefully we'll be able to use the worldspawn properties to "select"
the appropriate blocks. The max & min heights refer to separate
buildings. Hence, if a block is just one big building, then the max
and min heights will be perfectly aligned.

Also, added some more blocks. I think I might be reorganizing the
existing blocks at some point - possibly into height-bands, or
something like that.