Relabel the original blocks as "Series 01"
I want to keep better track of how many buildings fit certain
criteria. To know, off the top of my head, which buildings are packed
together tightly or loosely, to know the average height of a given
building.

I'm going to separate the blocks into "series" where each series
largely conforms in density and height characteristics. By virtue of
membership in these series, we'll be able to identify off-hand what
kind of block we're looking at.