Add some Voronoi block templates
For the first series of blocks, I was going to make a bunch of tightly
packed city blocks. But I'm no architectural genius. I can't just
whip up a bunch of tightly packed buildings out of the aether. I need
blueprints.

I can't generate something on the fly since we're doing handmade stuff
here. So I was racking my brain on how to make some sort of square
divsions when I remembered - VORONOI POLYGONS!

I generate up 15 voronoi polygon images using the "cell noise"
generator in GIMP (the GNU Image Manipulation Program). I used the
"Palletize" setting and set the "Shape" setting to 1.0, which resulted
in rounded rectanguloids - Manhattan Distance Voronoi Polygons. I then
colored these in and exported them as the bt_scXXX_seYYY texture
series. For each, "sc" refers to the scale setting in terms of 0.XXX.
The "se" refers to the seed use, purely of the form YYY. As you can
see, we made use of multiple sizes to (hopefully) simuate different
densities.

The first series to be based on this is s02_vrndense. Originally, it
was just s02 but that was too hard for me to pick out from everything
else. Anyway, this series sports an extra worldspawn field to specify
which Voronoi Polygon block template we used. The results are
promising so far - now I just need to get to work MAKING those blocks.