Sort BlockInspector's blocks alphabetically
For some reason the BlockInspector's different blocks were all over
the place. Up, down, differently stratified. I expected them to be in
alphabetical order but the order seemed... mostly random?

It was annoying enough that we now fill an array with our map-paths.
We then sort the array before copying these over to the choice-GUI.