Do a "shallow" copy of block meshes
Spawning in new meshes, building up new meshes - these things take
time! If we actually loaded a Qodot map for each block, the generator
would NEVER run fast enough.

Instead, we'll have all the meshes required loaded in, and we'll just
pass the meshes to facsimilie blocks. Of course, we needed to see
whether such a thing was even possible first!

The CopyBlock is the realization of this idea. It will be the base
unit of the city, and may even later feature additional randomization
on top of the city block. For right now, it's just a shell.