Crash Godot by loading too many maps (somehow)
Okay. So. Funny story: I was working on code to make-and-load all of
the derivative maps when the Phase3 main scene loads. That's nice and
logical - right?

Yeah I thought so too - except making all those derivatives and
loading them crashes Godot. And I don't mean it crashes my program - I
mean it hard crashes the engine. I had to hook in with GDB to try and
figure out what was going wrong. Something about not being able to
find one of the derivative .map files that DEFINITELY still exists.
Possibly an error with Qodot, but I'm gonna hold off judgement. I'm
REALLY pushing this plugin past what it was meant to do!

Although I'm nominally against pushing broken code to a git repo, I
felt that it was important to capture this particular moment because
this will have design ramifications for the rest of this project.
Maybe there's a way around this - perhaps if we tried loading them one
at a time, or otherwise deliberately staged?